requirements.txt
docling
python-dotenv
openai
python-dotenv
langchain
pgvector
psycopg2-binary
sqlalchemy>=1.4.0
fastapi
uvicorn
python-multipart
requests
numpy
